optimize vllm model (#4822)
* optimize varlen_fwd

* optimize add_rms_norm

* update UTs of varlen_fwd & add_rms_norm

* change another path of add_rms_norm to inplace

* support vllm api

* remove slot_mapping.int()

* optimize paged_attention api

* revert change in cpu part

* reformatted code

* fix CPU UT

* revert commit 33edeb6ee13431e8801531facfcc889a069776bd

* fix CPU UT

* add return to add_rms_norm

* add new api:single_query_memoized_kv_attention

* refine format

* refine UT

* add doc for params

* add single_query_memoized_kv_attention to cpu

* Update mha_fusion.py

* update doc and add head_mapping

* change single_query_memoized_kv_attention to single_query_kv_attention

---------

+++ csrc/gpu/aten/operators/xetla/kernels/SDP/fmha_forward_kernel.hpp
@@ -25,8 +25,8 @@ struct dispatch_fmha_forward_args_t {
   void* log_sumexp;
   float softmax_scale;
   float dropout_prob;
-  int32_t* cu_seqlen_q;
-  int32_t* cu_seqlen_k;
+  int64_t* cu_seqlen_q;
+  int64_t* cu_seqlen_k;
   uint32_t num_batches;
   uint32_t num_heads;
   uint32_t num_kv_heads;
